Configure load-tc358743-edid service to restart on failure
In testing TinyPilot on Debian Bullseye (Debian 11), I found that the load-tc358743-edid service fails at boot time, which causes uStreamer's capture to fail.

The ideal solution would be to find the service that needs to be up before the load-tc358743-edid service can succeed and then declare that in the After= line, but I haven't been able to identify which service that is.

As a workaround, we can add a retry count of 20 so that if we fail initially, we keep trying to start the service until it succeeds.