Drop support for Debian 9
We didn't advertise support for it, but we were testing it in CI. This drops support for Debian 9 because no Janus package is available on Debian 9, and there are no known Debian 9 consumers of this role.

We also adjust the meta/main.yml to accurately reflect support for both buster and bullseye versions of Debian (10, 11).