Update EDID "Product ID" to resolve blank screen issue in macOS 13.3+ (#101)
This updates the EDID in order to fix a compatibility issue which causes
blank screens on macOS Ventura 13.3 and newer.

## Details

We had several reports that macOS 13.3 and newer display as a blank
screen in TinyPilot. Our investigation revealed that this was due to
unexplained incompatibility with the "Product ID" and/or "Manufacturer"
name embedded in our existing EDID.

The existing "Product ID" was set to "0x8888" and this change updates it
to "0x9876". This cannot be easily seen directly in the source code but
can be verified by using a tool such as `edid-decode` or by viewing the
EDID in "AW EDID Editor".

## Testing

The fix has been verified using an M1 Mac Mini running macOS Ventura
13.3.1. The test process was to verify that no image was displayed using
the existing EDID, then to change to the new EDID and verify an image
was displayed.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/ansible-role-ustreamer/101"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>